refactoring forms.py into directory forms for better readibility and clean code
